+++ hyperdbg/include/SDK/imports/user/HyperDbgLibImports.h
@@ -118,6 +118,9 @@ hyperdbg_u_connect_remote_debugger_using_named_pipe(const CHAR * named_pipe, BOO
 IMPORT_EXPORT_LIBHYPERDBG BOOLEAN
 hyperdbg_u_connect_current_debugger_using_com_port(const CHAR * port_name, DWORD baudrate);
 
+IMPORT_EXPORT_LIBHYPERDBG BOOLEAN
+hyperdbg_u_debug_close_remote_debugger();
+
 //
 // Miscalenous functions
 //